remove person tab
also, move SpacesDetailsWorkstations to WorkstationContainer and modify so it can take either personId or spaceId
