+++ Keas.Mvc/ClientApp/components/People/PersonDetails.tsx
@@ -8,7 +8,12 @@ import {
     ModalHeader,
 } from "reactstrap";
 import { IPerson } from "../../Types";
-import Person from "../Person";
+import AccessContainer from "../Access/AccessContainer";
+import BioContainer from "../Biographical/BioContainer";
+import EquipmentContainer from "../Equipment/EquipmentContainer";
+import HistoryContainer from "../History/HistoryContainer";
+import KeyContainer from "../Keys/KeyContainer";
+import WorkstationContainer from "../Workstations/WorkstationContainer";
 import EquipmentEditValues from "./EquipmentEditValues";
 
 
@@ -31,7 +36,12 @@ export default class PersonDetails extends React.Component<IProps, {}> {
                 <Modal isOpen={this.props.modal} toggle={this.props.closeModal} size="lg">
                     <ModalHeader>Details for {this.props.selectedPerson.user.name}</ModalHeader>
                     <ModalBody>
-                        <Person person={this.props.selectedPerson}/>
+                        <BioContainer person={this.props.selectedPerson} />
+                        <KeyContainer person={this.props.selectedPerson} />
+                        <EquipmentContainer person={this.props.selectedPerson} />
+                        <AccessContainer person={this.props.selectedPerson} />
+                        <WorkstationContainer personId={this.props.selectedPerson.id} />
+                        <HistoryContainer controller="people" id={this.props.selectedPerson.id} />
                     </ModalBody>
                     <ModalFooter>
                         <Button color="secondary" onClick={this.props.closeModal}>
